requirements.txt
Pillow==8.0.1
selenium==3.141.0
urllib3==1.26.2

